feat: inventory -> Ledger 변경
